Validation

See merge request razeeb/gctl_erp!105

+++ GCTL.Service/FieldServices/EmployeeAdvanced/EmployeeAdvancedService.cs
@@ -58,7 +58,7 @@ namespace GCTL.Service.FieldServices.EmployeeAdvanced
 
                 //empadvance.EmployeeAdvanceID = emp.EmployeeAdvanceID;
                 empadvance.JobID = emp.JobID;
-                empadvance.AmountRequested = emp.AmountRequested;
+                empadvance.AmountRequested = (decimal)emp.AmountRequested;
                 empadvance.StartDate = emp.StartDate.HasValue ? DateOnly.FromDateTime(emp.StartDate.Value) : null;
                 empadvance.EndDate = emp.EndDate.HasValue ? DateOnly.FromDateTime(emp.EndDate.Value) : null;
                 empadvance.LIP = emp.LIP;
@@ -453,11 +453,14 @@ namespace GCTL.Service.FieldServices.EmployeeAdvanced
                     var beforeEntity = JsonConvert.DeserializeObject<EmployeeAdvancedVM>(JsonConvert.SerializeObject(entity, JsonSettings.IgnoreReferenceLoop));
 
                     entity.JobID = emp.JobID;
-                    entity.AmountRequested = emp.AmountRequested;
+                    entity.AmountRequested = (decimal)emp.AmountRequested;
                     entity.StartDate = emp.StartDate.HasValue ? DateOnly.FromDateTime(emp.StartDate.Value) : null;
                     entity.EndDate = emp.EndDate.HasValue ? DateOnly.FromDateTime(emp.EndDate.Value) : null;
 
-                    entity.RequestedByUserID = emp.RequestedByUserID != null && emp.RequestedByUserID.Count > 0 ? emp.RequestedByUserID.FirstOrDefault() : null;
+                    entity.RequestedByUserID =
+    emp.RequestedByUserID != null && emp.RequestedByUserID.Count > 0
+        ? emp.RequestedByUserID.FirstOrDefault()
+        : null;
 
                     entity.ApprovedByUserID = emp.ApprovedByUserID;
